requirements.txt
# Core
python-dotenv
pyyaml
requests
openai
anthropic
setuptools<81
keyring
keyrings.alt

# MCP
mcp[cli]

# Browser automation
playwright

# Dev tooling
pip-audit
pytest
pytest-asyncio
mypy
ruff

# Gateway server
fastapi
uvicorn[standard]
aiosqlite
aiohttp

# External integrations
neo4j
py2neo
dateparser

# Discord bot
discord.py[voice]>=2.0

# Telegram bot
python-telegram-bot[job-queue]>=21.0

# Scheduler
apscheduler>=3.10

